Dgefsd

+++ Coockie.java
@@ -2,15 +2,15 @@ import greenfoot.*;
 
 public class Coockie extends Actor
 {
-    public static int cash = 01;
+    public static double cash = 01;
     public void act()
     {
         if(Greenfoot.mouseClicked(this)){
             cash++;
         }
-        getWorld().showText("Coockies "+cash, 50, 15);
+        getWorld().showText("Coockies "+ cash, 50, 15);
     }
-    public void coockie(int how){
+    public void coockie(double how){
         cash = cash+how;
     }
 }
